requirements.txt
# Core web framework + Lambda adapter
fastapi==0.135.3
mangum==0.21.0
starlette==1.0.0
uvicorn==0.44.0

# HTTP client
requests==2.33.1
urllib3==2.6.3
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.11

# Data validation
pydantic==2.12.5
pydantic-core==2.41.5
annotated-types==0.7.0
typing-extensions==4.15.0
typing-inspection==0.4.2

# Environment variable management
python-dotenv==1.2.2

# Async support (required by starlette)
anyio==4.13.0

# ASGI server utilities
h11==0.16.0
click==8.3.2

# AWS SDK (pre-installed in Lambda runtime, listed for local dev)
# boto3
